feat: apply basic styling
