+++ src/components/App.css
@@ -1,33 +1,34 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
-  pointer-events: none;
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+/* ------------- screen-size neutral designs ------------ */
+.app-container {
   display: flex;
   flex-direction: column;
-  align-items: center;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+  max-width: 90vw;
+  padding: 1rem 2rem 1rem 2rem;
 }
 
-.App-link {
-  color: #61dafb;
+.search-results-container {
+  display: flex;
+  flex-direction: column;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.card-container {
+  align-items: center;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+.card {
+  width: 15rem;
+  margin: 0.5rem;
+}
+
+.card img {
+  width: 15rem;
+}
+
+.card a {
+  color: #252525;
+  text-decoration: none;
 }
