+++ src/index.js
@@ -1,5 +1,3 @@
-import "./main.css";
-
 import React from "react";
 import ReactDOM from "react-dom";
